requirements.txt
uvicorn~=0.35.0
fastapi~=0.116.1
pillow~=10.2.0
pydantic~=2.11.7

python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
python-magic==0.4.27
aiofiles==23.2.1
